New tool definitions in ids.h. Left mouse button capability increased. Smoother lines added. Accuracy increased by using double instead of float in matrix operations.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6953 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/ltsgraph3d/ltsgraph3d.cpp
@@ -368,7 +368,7 @@ void LTSGraph3d::setTool(bool isColour)
   colouring = isColour;
 }
 
-void LTSGraph3d::getCanvasMdlvwMtrx(float * mtrx)
+void LTSGraph3d::getCanvasMdlvwMtrx(double * mtrx)
 {
   glCanvas->getMdlvwMtrx(mtrx);
 }
